# openapi.yaml
...
openapi: 3.1.0
info:
  title: SMIRK API
  version: "1.0.0"
  description: Route inventory generated from concrete Express route declarations in server.ts and src/team-routes.ts. Middleware app.use declarations are excluded from paths.
servers:
  - url: https://smirkcalls.com
    description: Production
  - url: http://localhost:3000
    description: Local development
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
    WorkspaceBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Workspace session or invite token
    ProvisioningBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Provisioning token
paths:
...
  "/api/recovery/{callSid}/call-back":
    post:
      operationId: "POST_api_recovery_callSid_call_back"
      tags:
        - "recovery"
      summary: "POST /api/recovery/:callSid/call-back"
      x-source: "src/routes/recovery-routes.ts:124"
      x-express-path: "/api/recovery/:callSid/call-back"
      security:
        - WorkspaceBearerAuth: []
        - ApiKeyAuth: []
      responses:
        "200":
          description: Success response. Body shape is implemented in the route handler.
        "400":
          description: Bad request or validation failure.
        "401":
          description: Authentication required or invalid credentials.
        "500":
          description: Server error.
  "/api/recovery/{callSid}/close":
    post:
      operationId: "POST_api_recovery_callSid_close"
      tags:
        - "recovery"
      summary: "POST /api/recovery/:callSid/close"
      x-source: "src/routes/recovery-routes.ts:169"
      x-express-path: "/api/recovery/:callSid/close"
      security:
        - WorkspaceBearerAuth: []
        - ApiKeyAuth: []
      responses:
        "200":
          description: Success response. Body shape is implemented in the route handler.
        "400":
          description: Bad request or validation failure.
        "401":
          description: Authentication required or invalid credentials.
        "500":
          description: Server error.
  "/api/recovery/direct-dial":
    post:
      operationId: "POST_api_recovery_direct_dial"
      tags:
        - "recovery"
      summary: "POST /api/recovery/direct-dial"
      x-source: "src/routes/recovery-routes.ts:190"
      x-express-path: "/api/recovery/direct-dial"
      security:
        - ApiKeyAuth: []
      responses:
        "200":
          description: Success response. Body shape is implemented in the route handler.
        "400":
          description: Bad request or validation failure.
        "401":
          description: Authentication required or invalid credentials.
        "500":
          description: Server error.
  "/api/recovery/queue":
    get:
      operationId: "GET_api_recovery_queue"
      tags:
        - "recovery"
      summary: "GET /api/recovery/queue"
      x-source: "src/routes/recovery-routes.ts:36"
      x-express-path: "/api/recovery/queue"
      security:
        - WorkspaceBearerAuth: []
        - ApiKeyAuth: []
      responses:
        "200":
          description: Success response. Body shape is implemented in the route handler.
        "400":
          description: Bad request or validation failure.
        "401":
          description: Authentication required or invalid credentials.
        "500":
          description: Server error.
  "/api/recovery/stats":
    get:
      operationId: "GET_api_recovery_stats"
      tags:
        - "recovery"
      summary: "GET /api/recovery/stats"
      x-source: "src/routes/recovery-routes.ts:218"
      x-express-path: "/api/recovery/stats"
      security:
        - WorkspaceBearerAuth: []
        - ApiKeyAuth: []
      responses:
        "200":
          description: Success response. Body shape is implemented in the route handler.
        "400":
          description: Bad request or validation failure.
        "401":
          description: Authentication required or invalid credentials.
        "500":
          description: Server error.
...